Reorganize util code in ansible-test.
Code in util.py that depends on CommonConfig is now in util_common.py.
