+++ test/runner/lib/docker_util.py
@@ -6,16 +6,16 @@ import json
 import os
 import time
 
-from lib.executor import (
-    SubprocessError,
-)
-
 from lib.util import (
     ApplicationError,
-    run_command,
     common_environment,
     display,
     find_executable,
+    SubprocessError,
+)
+
+from lib.util_common import (
+    run_command,
 )
 
 from lib.config import (
