+++ test/runner/lib/sanity/ansible_doc.py
@@ -16,8 +16,11 @@ from lib.sanity import (
 from lib.util import (
     SubprocessError,
     display,
-    intercept_command,
     read_lines_without_comments,
+)
+
+from lib.util_common import (
+    intercept_command,
 )
 
 from lib.ansible_util import (
