+++ test/runner/lib/sanity/compile.py
@@ -13,12 +13,15 @@ from lib.sanity import (
 
 from lib.util import (
     SubprocessError,
-    run_command,
     display,
     find_python,
     read_lines_without_comments,
     parse_to_list_of_dict,
     INSTALL_ROOT,
+)
+
+from lib.util_common import (
+    run_command,
 )
 
 from lib.config import (
