+++ test/runner/lib/sanity/rstcheck.py
@@ -13,11 +13,14 @@ from lib.sanity import (
 
 from lib.util import (
     SubprocessError,
-    run_command,
     parse_to_list_of_dict,
     display,
     read_lines_without_comments,
     INSTALL_ROOT,
+)
+
+from lib.util_common import (
+    run_command,
 )
 
 from lib.config import (
