+++ test/runner/lib/sanity/shellcheck.py
@@ -18,8 +18,11 @@ from lib.sanity import (
 
 from lib.util import (
     SubprocessError,
-    run_command,
     read_lines_without_comments,
+)
+
+from lib.util_common import (
+    run_command,
 )
 
 from lib.config import (
